Add initial project structure with batch processing capabilities for DeepSeek OCR
- Created .gitignore to exclude virtual environment and environment files.
- Added .python-version to specify Python version.
- Implemented main.py with a simple entry point.
- Introduced BATCH_PROCESSING_README.md for batch processing instructions.
- Added configuration and Dagster integration for batch processing in dagster_deepseek_ocr.py.
- Developed example usage script for batch processing.
- Implemented core batch processing logic in run_dpsk_ocr_pdf_batch.py.
- Added support for processing multiple PDFs with detailed output structure.
- Included image processing and bounding box drawing functionalities.
- Updated config.py with paths for input and output.
- Enhanced error handling and logging throughout the processing pipeline.
- Added support for concurrent processing of PDFs.

+++ DeepSeek-OCR-master/DeepSeek-OCR-vllm/deepencoder/sam_vary_sdpa.py
@@ -472,7 +472,7 @@ class PatchEmbed(nn.Module):
         )
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
-        x = self.proj(x)
+        x = self.proj(x.to(self.proj.weight.dtype))
         # B C H W -> B H W C
         x = x.permute(0, 2, 3, 1)
         return x
